ImageModel/requirements.txt
# Core ML and Data Science
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==1.7.6

# Deep Learning Frameworks
torch
torchvision
tensorflow

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Computer Vision
opencv-python==4.9.0.80
Pillow==10.2.0
matplotlib==3.9.2

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.20.0
streamlit==1.36.0

# Data Formats
openpyxl==3.1.5
pyarrow==15.0.2
h5py==3.11.0

# Model Management
joblib==1.4.2

# Optimization and Experiment Tracking
optuna==3.6.1
mlflow==2.14.1
ultralytics==8.2.30

# Development
black==24.4.2
flake8==7.1.1
